Moved files around. Cleaned up a little.
