+++ data/morphology.ts
@@ -1,75 +1,5 @@
-const bookFiles = [
-  'Gen.json',
-  'Exo.json',
-  'Lev.json',
-  'Num.json',
-  'Deu.json',
-  'Jos.json',
-  'Jdg.json',
-  'Rut.json',
-  '1Sa.json',
-  '2Sa.json',
-  '1Ki.json',
-  '2Ki.json',
-  '1Ch.json',
-  '2Ch.json',
-  'Ezr.json',
-  'Neh.json',
-  'Est.json',
-  'Job.json',
-  'Psa.json',
-  'Pro.json',
-  'Ecc.json',
-  'Sng.json',
-  'Isa.json',
-  'Jer.json',
-  'Lam.json',
-  'Ezk.json',
-  'Dan.json',
-  'Hos.json',
-  'Jol.json',
-  'Amo.json',
-  'Oba.json',
-  'Jon.json',
-  'Mic.json',
-  'Nam.json',
-  'Hab.json',
-  'Zep.json',
-  'Hag.json',
-  'Zec.json',
-  'Mal.json',
-  'Mat.json',
-  'Mrk.json',
-  'Luk.json',
-  'Jhn.json',
-  'Act.json',
-  'Rom.json',
-  '1Co.json',
-  '2Co.json',
-  'Gal.json',
-  'Eph.json',
-  'Php.json',
-  'Col.json',
-  '1Th.json',
-  '2Th.json',
-  '1Ti.json',
-  '2Ti.json',
-  'Tit.json',
-  'Phm.json',
-  'Heb.json',
-  'Jas.json',
-  '1Pe.json',
-  '2Pe.json',
-  '1Jn.json',
-  '2Jn.json',
-  '3Jn.json',
-  'Jud.json',
-  'Rev.json',
-];
+import { bookKeys } from './book-keys';
 
-export default Object.fromEntries(
-  bookFiles.map((fileName) => [
-    fileName.replace('.json', ''),
-    require(`./morphology/${fileName}`) as string[][][],
-  ])
+export const morphologyData = Object.fromEntries(
+  bookKeys.map((key) => [require(`./morphology/${key}`) as string[][][]])
 );
