+++ docs/i18n.md
@@ -19,7 +19,7 @@ The data in `packages/web/src/public/locales/<locale-code>/bible.json` is used t
 
 The first line is the title of the book (Genesis). The second line is the short-hand abbreviation of the book title (Gen).
 
-The first three characters of each key in `bible.json` identify the book. See `packages/web/src/features/translation/book-keys.ts` for an array of all the book keys. If you copied `bible.json` from another language, you should already have all the correct keys.
+The first three characters of each key in `bible.json` identify the book. See `data/book-keys.ts` for an array of all the book keys. You should use the lowercase version of each key If you copied `bible.json` from another language, you should already have all the correct keys.
 
 The `reference_format` key at the top of the file is for formatting references and parsing references.
