+++ packages/api-types/src/index.ts
@@ -1,5 +1,3 @@
-export { bookKeys } from './book-keys';
-
 // Prisma handles enums differently according to this issue in typescript:
 // https://github.com/microsoft/TypeScript/issues/3192#issuecomment-261720275
 function makeEnum<T extends { [index: string]: U }, U extends string>(x: T) {
